Huge refactor - pull out shared logic into struct in a new lib, add plugin trait
