+++ rust/sealevel/Cargo.toml
@@ -11,4 +11,5 @@ members = [
     "libraries/message-recipient-interface",
     "libraries/multisig-ism",
     "libraries/serializable-account-meta",
+    "libraries/hyperlane-sealevel-token",
 ]
