+++ rust/sealevel/libraries/hyperlane-sealevel-token/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "hyperlane-sealevel-token"
+name = "hyperlane-sealevel-token-lib"
 version = "0.1.0"
 edition = "2021"
 
@@ -9,9 +9,8 @@ no-entrypoint = []
 [dependencies]
 bs58 = "=0.4.0" # FIXME remove
 hyperlane-core = { path = "../../hyperlane-core" }
-hyperlane-sealevel-token-lib = { path = "../../libraries/hyperlane-sealevel-token" }
-hyperlane-sealevel-mailbox = { path = "../mailbox", features = ["no-entrypoint"] }
-hyperlane-sealevel-message-recipient-interface = { path = "../../libraries/message-recipient-interface" }
+hyperlane-sealevel-mailbox = { path = "../../programs/mailbox", features = ["no-entrypoint"] }
+hyperlane-sealevel-message-recipient-interface = { path = "../message-recipient-interface" }
 borsh = "0.9"
 num-derive = "0.3"
 num-traits = "0.2"
