+++ rust/sealevel/libraries/hyperlane-sealevel-token/Xargo.toml
@@ -1,0 +1,3 @@
+# FIXME do we need this?
+[target.bpfel-unknown-unknown.dependencies.std]
+features = []
