+++ rust/sealevel/programs/hyperlane-sealevel-token/src/processor.rs
@@ -1,40 +1,17 @@
 //! TODO
 
 use borsh::BorshSerialize;
-use hyperlane_core::{Decode, Encode as _, H256};
-use hyperlane_sealevel_mailbox::{
-    instruction::{Instruction as MailboxIxn, OutboxDispatch as MailboxOutboxDispatch},
-    mailbox_outbox_pda_seeds,
-};
 use hyperlane_sealevel_message_recipient_interface::MessageRecipientInstruction;
+use hyperlane_sealevel_token_lib::{
+    instruction::{Init, TransferFromRemote, TransferRemote},
+    processor::HyperlaneSealevelToken,
+};
 use solana_program::{
-    account_info::{next_account_info, AccountInfo},
-    entrypoint,
-    entrypoint::ProgramResult,
-    instruction::{AccountMeta, Instruction},
-    msg,
-    program::{invoke, invoke_signed, set_return_data},
-    program_error::ProgramError,
-    program_pack::Pack as _,
-    pubkey::Pubkey,
-    rent::Rent,
-    system_instruction,
+    account_info::AccountInfo, entrypoint, entrypoint::ProgramResult, msg,
+    program::set_return_data, program_error::ProgramError, pubkey::Pubkey,
 };
-use spl_associated_token_account::{
-    get_associated_token_address_with_program_id,
-    instruction::create_associated_token_account_idempotent,
-};
-use spl_token_2022::instruction::{burn_checked, mint_to_checked};
 
-use crate::{
-    accounts::{HyperlaneToken, HyperlaneTokenAccount},
-    error::Error,
-    instruction::{
-        Event, EventReceivedTransferRemote, EventSentTransferRemote, Init, Instruction as TokenIxn,
-        TransferFromRemote, TransferRemote,
-    },
-    message::TokenMessage,
-};
+use crate::{instruction::Instruction as TokenIxn, plugin::SyntheticPlugin};
 
 #[cfg(not(feature = "no-entrypoint"))]
 entrypoint!(process_instruction);
@@ -50,37 +27,6 @@ macro_rules! hyperlane_token_pda_seeds {
         &[b"hyperlane_token", b"-", b"token", &[$bump_seed]]
     }};
 }
-
-/// Seeds relating to the PDA account that acts both as the mint
-/// *and* the mint authority.
-#[macro_export]
-macro_rules! hyperlane_token_mint_pda_seeds {
-    () => {{
-        &[b"hyperlane_token", b"-", b"mint"]
-    }};
-
-    ($bump_seed:expr) => {{
-        &[b"hyperlane_token", b"-", b"mint", &[$bump_seed]]
-    }};
-}
-
-#[macro_export]
-macro_rules! hyperlane_token_ata_payer_pda_seeds {
-    () => {{
-        &[b"hyperlane_token", b"-", b"ata_payer"]
-    }};
-
-    ($bump_seed:expr) => {{
-        &[b"hyperlane_token", b"-", b"ata_payer", &[$bump_seed]]
-    }};
-}
-
-// TODO make these easily configurable?
-pub const REMOTE_DECIMALS: u8 = 18;
-pub const DECIMALS: u8 = 8;
-
-/// The size of the mint account in bytes.
-const MINT_ACCOUNT_SIZE: usize = spl_token_2022::state::Mint::LEN;
 
 pub fn process_instruction(
     program_id: &Pubkey,
@@ -142,107 +88,7 @@ pub fn process_instruction(
 /// 3. [writable] The ATA payer PDA account.
 /// 4. [signer] The payer.
 fn initialize(program_id: &Pubkey, accounts: &[AccountInfo], init: Init) -> ProgramResult {
-    // On chain create appears to use realloc which is limited to 1024 byte increments.
-    let token_account_size = 2048;
-
-    let accounts_iter = &mut accounts.iter();
-
-    // Account 0: System program
-    let system_program = next_account_info(accounts_iter)?;
-    if system_program.key != &solana_program::system_program::id() {
-        return Err(ProgramError::InvalidArgument);
-    }
-
-    // Account 1: Token storage account
-    let token_account = next_account_info(accounts_iter)?;
-    let (token_key, token_bump) =
-        Pubkey::find_program_address(hyperlane_token_pda_seeds!(), program_id);
-    if &token_key != token_account.key {
-        return Err(ProgramError::InvalidArgument);
-    }
-
-    // Account 2: Mint / mint authority
-    let mint_account = next_account_info(accounts_iter)?;
-    let (mint_key, mint_bump) =
-        Pubkey::find_program_address(hyperlane_token_mint_pda_seeds!(), program_id);
-    if &mint_key != mint_account.key {
-        return Err(ProgramError::InvalidArgument);
-    }
-
-    // Account 3: ATA payer.
-    let ata_payer_account = next_account_info(accounts_iter)?;
-    let (ata_payer_key, ata_payer_bump) =
-        Pubkey::find_program_address(hyperlane_token_ata_payer_pda_seeds!(), program_id);
-    if &ata_payer_key != ata_payer_account.key {
-        return Err(ProgramError::InvalidArgument);
-    }
-
-    // Account 4: Payer
-    let payer_account = next_account_info(accounts_iter)?;
-    if !payer_account.is_signer {
-        return Err(ProgramError::MissingRequiredSignature);
-    }
-
-    if accounts_iter.next().is_some() {
-        return Err(ProgramError::from(Error::ExtraneousAccount));
-    }
-
-    // Create the ATA payer.
-    // This is a separate PDA because the ATA program requires
-    // the payer to have no data in it.
-    invoke_signed(
-        &system_instruction::create_account(
-            payer_account.key,
-            ata_payer_account.key,
-            Rent::default().minimum_balance(0),
-            0,
-            // Grant ownership to the system program so that the ATA program
-            // can call into the system program with the ATA payer as the
-            // payer.
-            &solana_program::system_program::id(),
-        ),
-        &[payer_account.clone(), ata_payer_account.clone()],
-        &[hyperlane_token_ata_payer_pda_seeds!(ata_payer_bump)],
-    )?;
-
-    // Create token account PDA
-    invoke_signed(
-        &system_instruction::create_account(
-            payer_account.key,
-            token_account.key,
-            Rent::default().minimum_balance(token_account_size),
-            token_account_size.try_into().unwrap(),
-            program_id,
-        ),
-        &[payer_account.clone(), token_account.clone()],
-        &[hyperlane_token_pda_seeds!(token_bump)],
-    )?;
-
-    let token = HyperlaneToken {
-        bump: token_bump,
-        mailbox: init.mailbox,
-        mailbox_local_domain: init.mailbox_local_domain,
-        mint: mint_key,
-        mint_bump: mint_bump,
-        ata_payer_bump,
-    };
-    HyperlaneTokenAccount::from(token).store(token_account, true)?;
-
-    // Create mint / mint authority PDA.
-    // Grant ownership to the SPL token program.
-    invoke_signed(
-        &system_instruction::create_account(
-            payer_account.key,
-            mint_account.key,
-            Rent::default().minimum_balance(MINT_ACCOUNT_SIZE),
-            MINT_ACCOUNT_SIZE.try_into().unwrap(),
-            &spl_token_2022::id(),
-        ),
-        &[payer_account.clone(), mint_account.clone()],
-        &[hyperlane_token_mint_pda_seeds!(mint_bump)],
-    )?;
-
-    Ok(())
+    HyperlaneSealevelToken::<SyntheticPlugin>::initialize(program_id, accounts, init)
 }
 
 /// Transfers tokens to a remote.
@@ -261,156 +107,9 @@ fn initialize(program_id: &Pubkey, accounts: &[AccountInfo], init: Init) -> Prog
 fn transfer_remote(
     program_id: &Pubkey,
     accounts: &[AccountInfo],
-    xfer: TransferRemote,
+    transfer: TransferRemote,
 ) -> ProgramResult {
-    let amount: u64 = xfer.amount_or_id.try_into().map_err(|_| Error::TODO)?;
-
-    let accounts_iter = &mut accounts.iter();
-
-    // Account 0: SPL Noop
-    let spl_noop = next_account_info(accounts_iter)?;
-    if spl_noop.key != &spl_noop::id() || !spl_noop.executable {
-        return Err(ProgramError::InvalidArgument);
-    }
-
-    // Account 1: Token storage account
-    let token_account = next_account_info(accounts_iter)?;
-    let token =
-        HyperlaneTokenAccount::fetch(&mut &token_account.data.borrow_mut()[..])?.into_inner();
-    let token_seeds: &[&[u8]] = hyperlane_token_pda_seeds!(token.bump);
-    let expected_token_key = Pubkey::create_program_address(token_seeds, program_id)?;
-    if token_account.key != &expected_token_key {
-        return Err(ProgramError::InvalidArgument);
-    }
-    if token_account.owner != program_id {
-        return Err(ProgramError::IncorrectProgramId);
-    }
-
-    // Account 2: Mailbox program
-    let mailbox_info = next_account_info(accounts_iter)?;
-    if mailbox_info.key != &token.mailbox {
-        return Err(ProgramError::IncorrectProgramId);
-    }
-    // TODO supposed to use create_program_address() but we would need to pass in bump seed...
-
-    // Account 3: Mailbox outbox data account
-    // TODO should I be using find_program_address...?
-    let mailbox_outbox_account = next_account_info(accounts_iter)?;
-    let (mailbox_outbox, _mailbox_outbox_bump) = Pubkey::find_program_address(
-        mailbox_outbox_pda_seeds!(token.mailbox_local_domain),
-        &token.mailbox,
-    );
-    if mailbox_outbox_account.key != &mailbox_outbox {
-        return Err(ProgramError::InvalidArgument);
-    }
-
-    // Account 4: Sender account
-    let sender_wallet = next_account_info(accounts_iter)?;
-    if !sender_wallet.is_signer {
-        return Err(ProgramError::MissingRequiredSignature);
-    }
-
-    // 5. SPL token 2022 program
-    let spl_token_2022 = next_account_info(accounts_iter)?;
-    if spl_token_2022.key != &spl_token_2022::id() || !spl_token_2022.executable {
-        return Err(ProgramError::InvalidArgument);
-    }
-
-    // 6. mint account
-    let mint_account = next_account_info(accounts_iter)?;
-    let mint_seeds: &[&[u8]] = hyperlane_token_mint_pda_seeds!(token.mint_bump);
-    let expected_mint_key = Pubkey::create_program_address(mint_seeds, program_id)?;
-    if mint_account.key != &expected_mint_key {
-        return Err(ProgramError::InvalidArgument);
-    }
-    if *mint_account.key != token.mint {
-        return Err(ProgramError::InvalidArgument);
-    }
-    if mint_account.owner != &spl_token_2022::id() {
-        return Err(ProgramError::IncorrectProgramId);
-    }
-    if mint_account.owner != &spl_token_2022::id() {
-        return Err(ProgramError::IncorrectProgramId);
-    }
-
-    // Hmmmmmm should this be enforced?
-
-    // 7. sender associated token account
-    let sender_ata = next_account_info(accounts_iter)?;
-    let expected_sender_associated_token_account = get_associated_token_address_with_program_id(
-        sender_wallet.key,
-        mint_account.key,
-        &spl_token_2022::id(),
-    );
-    if sender_ata.key != &expected_sender_associated_token_account {
-        return Err(ProgramError::InvalidArgument);
-    }
-    if accounts_iter.next().is_some() {
-        return Err(ProgramError::from(Error::ExtraneousAccount));
-    }
-
-    let burn_ixn = burn_checked(
-        &spl_token_2022::id(),
-        sender_ata.key,
-        mint_account.key,
-        sender_wallet.key,
-        &[sender_wallet.key],
-        amount,
-        DECIMALS,
-    )?;
-    // Sender wallet is expected to have signed this transaction
-    invoke(
-        &burn_ixn,
-        &[
-            sender_ata.clone(),
-            mint_account.clone(),
-            sender_wallet.clone(),
-        ],
-    )?;
-
-    let token_xfer_message =
-        TokenMessage::new_erc20(xfer.recipient, xfer.amount_or_id, vec![]).to_vec();
-    let mailbox_ixn = MailboxIxn::OutboxDispatch(MailboxOutboxDispatch {
-        sender: *token_account.key,
-        local_domain: token.mailbox_local_domain,
-        destination_domain: xfer.destination_domain,
-        recipient: xfer.destination_program_id,
-        message_body: token_xfer_message,
-    });
-    let mailbox_ixn = Instruction {
-        program_id: token.mailbox,
-        data: mailbox_ixn.into_instruction_data().unwrap(),
-        accounts: vec![
-            AccountMeta::new(*mailbox_outbox_account.key, false),
-            AccountMeta::new_readonly(*token_account.key, true),
-            AccountMeta::new_readonly(spl_noop::id(), false),
-        ],
-    };
-    // TODO implement interchain gas payment via paymaster? dispatch_with_gas()?
-    invoke_signed(
-        &mailbox_ixn,
-        &[
-            mailbox_outbox_account.clone(),
-            token_account.clone(),
-            spl_noop.clone(),
-        ],
-        &[token_seeds],
-    )?;
-
-    let event = Event::new(EventSentTransferRemote {
-        destination: xfer.destination_domain,
-        recipient: xfer.recipient,
-        amount: xfer.amount_or_id,
-    });
-    let event_data = event.to_noop_cpi_ixn_data().map_err(|_| Error::TODO)?;
-    let noop_cpi_log = Instruction {
-        program_id: spl_noop::id(),
-        accounts: vec![],
-        data: event_data,
-    };
-    invoke_signed(&noop_cpi_log, &[], &[token_seeds])?;
-
-    Ok(())
+    HyperlaneSealevelToken::<SyntheticPlugin>::transfer_remote(program_id, accounts, transfer)
 }
 
 // Accounts:
@@ -427,174 +126,7 @@ fn transfer_remote(
 fn transfer_from_remote(
     program_id: &Pubkey,
     accounts: &[AccountInfo],
-    xfer: TransferFromRemote,
+    transfer: TransferFromRemote,
 ) -> ProgramResult {
-    let mut message_reader = std::io::Cursor::new(xfer.message);
-    let message = TokenMessage::read_from(&mut message_reader)
-        .map_err(|_err| ProgramError::from(Error::TODO))?;
-    // FIXME we must account for decimals of the mint not only the raw amount value during
-    // transfer. Wormhole accounts for this with some extra care taken to round/truncate properly -
-    // we should do the same.
-    let amount = message.amount().try_into().map_err(|_| Error::TODO)?;
-    // FIXME validate message fields?
-
-    let accounts_iter = &mut accounts.iter();
-
-    // FIXME validate mailbox auth pda and require that it's a signer
-    // Account 0: Mailbox authority
-    let _mailbox_auth = next_account_info(accounts_iter)?;
-
-    // Account 1: System program
-    let system_program = next_account_info(accounts_iter)?;
-    if system_program.key != &solana_program::system_program::id() {
-        return Err(ProgramError::InvalidArgument);
-    }
-    // Account 2: SPL Noop program
-    let spl_noop = next_account_info(accounts_iter)?;
-    if spl_noop.key != &spl_noop::id() || !spl_noop.executable {
-        return Err(ProgramError::InvalidArgument);
-    }
-
-    // Account 3: Token account
-    let token_account = next_account_info(accounts_iter)?;
-    let token =
-        HyperlaneTokenAccount::fetch(&mut &token_account.data.borrow_mut()[..])?.into_inner();
-    let token_seeds: &[&[u8]] = hyperlane_token_pda_seeds!(token.bump);
-    let expected_token_key = Pubkey::create_program_address(token_seeds, program_id)?;
-    if token_account.key != &expected_token_key {
-        return Err(ProgramError::InvalidArgument);
-    }
-    if token_account.owner != program_id {
-        return Err(ProgramError::IncorrectProgramId);
-    }
-
-    // Account 4: Recipient wallet
-    let recipient_wallet = next_account_info(accounts_iter)?;
-
-    // Account 5: SPL token 2022 program
-    let spl_token_2022 = next_account_info(accounts_iter)?;
-    if spl_token_2022.key != &spl_token_2022::id() || !spl_token_2022.executable {
-        return Err(ProgramError::InvalidArgument);
-    }
-    // Account 6: SPL associated token account
-    let spl_ata = next_account_info(accounts_iter)?;
-    if spl_ata.key != &spl_associated_token_account::id() || !spl_ata.executable {
-        return Err(ProgramError::InvalidArgument);
-    }
-
-    // Account 7: Mint account
-    let mint_account = next_account_info(accounts_iter)?;
-    let mint_seeds: &[&[u8]] = hyperlane_token_mint_pda_seeds!(token.mint_bump);
-    let expected_mint_key = Pubkey::create_program_address(mint_seeds, program_id)?;
-    if mint_account.key != &expected_mint_key {
-        return Err(ProgramError::InvalidArgument);
-    }
-    if mint_account.owner != &spl_token_2022::id() {
-        return Err(ProgramError::IncorrectProgramId);
-    }
-
-    // Account 8: Recipient associated token account
-    let recipient_ata = next_account_info(accounts_iter)?;
-    let expected_recipient_associated_token_account = get_associated_token_address_with_program_id(
-        recipient_wallet.key,
-        mint_account.key,
-        &spl_token_2022::id(),
-    );
-    if recipient_ata.key != &expected_recipient_associated_token_account {
-        return Err(ProgramError::InvalidArgument);
-    }
-
-    // Account 9: ATA payer PDA account
-    let ata_payer_account = next_account_info(accounts_iter)?;
-    let ata_payer_seeds: &[&[u8]] = hyperlane_token_ata_payer_pda_seeds!(token.ata_payer_bump);
-    let expected_ata_payer_account = Pubkey::create_program_address(ata_payer_seeds, program_id)?;
-    if ata_payer_account.key != &expected_ata_payer_account {
-        return Err(ProgramError::InvalidArgument);
-    }
-
-    if accounts_iter.next().is_some() {
-        return Err(ProgramError::from(Error::ExtraneousAccount));
-    }
-
-    // Create and init (this does both) associated token account if necessary.
-    invoke_signed(
-        &create_associated_token_account_idempotent(
-            ata_payer_account.key,
-            // payer_account.key,
-            recipient_wallet.key,
-            mint_account.key,
-            &spl_token_2022::id(),
-        ),
-        &[
-            ata_payer_account.clone(),
-            // token_account.clone(),
-            // mint_account.clone(),
-            // solana_program::instruction::AccountMeta::new_readonly(*mint_account.key, true),
-            // payer_account.clone(),
-            recipient_ata.clone(),
-            recipient_wallet.clone(),
-            mint_account.clone(),
-            system_program.clone(),
-            spl_token_2022.clone(),
-        ],
-        &[ata_payer_seeds, mint_seeds],
-    )?;
-
-    // After potentially paying for the ATA creation, we need to make sure
-    // the ATA payer still meets the rent-exemption requirements!
-    let ata_payer_lamports = ata_payer_account.lamports();
-    let ata_payer_rent_exemption_requirement = Rent::default().minimum_balance(0);
-    if ata_payer_lamports < ata_payer_rent_exemption_requirement {
-        return Err(ProgramError::from(Error::AtaBalanceTooLow));
-    }
-
-    // Mints new tokens to an account.  The native mint does not support
-    // minting.
-    //
-    // Accounts expected by this instruction:
-    //
-    //   * Single authority
-    //   0. `[writable]` The mint.
-    //   1. `[writable]` The account to mint tokens to.
-    //   2. `[signer]` The mint's minting authority.
-    //
-    //   * Multisignature authority
-    //   0. `[writable]` The mint.
-    //   1. `[writable]` The account to mint tokens to.
-    //   2. `[]` The mint's multisignature mint-tokens authority.
-    //   3. ..3+M `[signer]` M signer accounts.
-    let mint_ixn = mint_to_checked(
-        &spl_token_2022::id(),
-        mint_account.key,
-        recipient_ata.key,
-        mint_account.key,
-        &[],
-        amount,
-        DECIMALS,
-    )?;
-    invoke_signed(
-        &mint_ixn,
-        &[
-            mint_account.clone(),
-            recipient_ata.clone(),
-            mint_account.clone(),
-        ],
-        &[hyperlane_token_mint_pda_seeds!(token.mint_bump)],
-    )?;
-
-    let event = Event::new(EventReceivedTransferRemote {
-        origin: xfer.origin,
-        // Note: assuming recipient not recipient ata is the correct "recipient" to log.
-        recipient: H256::from(recipient_wallet.key.to_bytes()),
-        amount: message.amount(),
-    });
-    let event_data = event.to_noop_cpi_ixn_data().map_err(|_| Error::TODO)?;
-    let noop_cpi_log = Instruction {
-        program_id: spl_noop::id(),
-        accounts: vec![],
-        data: event_data,
-    };
-    invoke_signed(&noop_cpi_log, &[], &[token_seeds])?;
-
-    Ok(())
+    HyperlaneSealevelToken::<SyntheticPlugin>::transfer_from_remote(program_id, accounts, transfer)
 }
